agent code segregation
